+++ src/PointToCursorHandleConverter.ts
@@ -1,26 +1,19 @@
 import { assert, minByDirectedWalker, getDepth, mapComparer, Comparer, walkAround } from './utils';
 import { ManhattenComparerToBoundary } from './ManhattanToBoundaryComparer';
+import { HorizontalClosestDistanceType, ManhattanOffset, MinDistances } from './MinDistances';
 
-export const LOCATION_ATTR_NAME = "data-loc";
-export type Point = { x: number, y: number };
-export type Distance = { dx: number, dy: number };
+export const LOCATION_ATTR_NAME = 'data-loc';
+export type Point = { x: number; y: number };
+export type Distance = { dx: number; dy: number };
 
-
-export type ManhattanOffset = { offsetFromLeft: number, offsetFromRight: number, offsetFromTop: number, offsetFromBottom: number };
-// "offset" means w.r.t. the positve axis, e.g.:
-// - a negative left offset means the point is to the left of the rect.left
-// - a negative right offset means the point is to the left of rect.right
-// It means that the offset can be regarded as one of the dimensions of the offset vector that, when added to a boundary coordinate, ends up at the the point (at least that location).
-
-export type SourceLocation = { start: number, end: number };
+export type SourceLocation = { start: number; end: number };
 
 export function getCursorIndexByProximity(element: HTMLElement, point: Point): number | undefined {
-    type T = { distances: MinDistances, loc: SourceLocation, depth: number };
+    type T = { distances: MinDistances; loc: SourceLocation; depth: number };
 
     function select(e: HTMLElement): T | undefined {
         const loc = selectLocation(e);
-        if (loc === undefined)
-            return undefined;
+        if (loc === undefined) return undefined;
         const distances = MinDistances.fromManhattan(getDistance(e, point));
 
         return { distances, loc, depth: getDepth(e) };
@@ -30,10 +23,9 @@ export function getCursorIndexByProximity(element: HTMLElement, point: Point): n
     const bests = minByDirectedWalker<T>(element, select, mapComparerToDistances(comparer));
 
     // @ts-ignore
-    const datalocsDebug = bests.map(b => b.element.attributes["data-loc"].value); // eslint-disable-line @typescript-eslint/no-unused-vars
+    const datalocsDebug = bests.map(b => b.element.attributes['data-loc'].value); // eslint-disable-line @typescript-eslint/no-unused-vars
 
-    if (bests.length === 0)
-        return undefined;
+    if (bests.length === 0) return undefined;
     const best = bests[0];
     const result = apply(best.value.distances, best.value.loc);
     return result;
@@ -52,27 +44,24 @@ function compareByMinHorizontalDistance(a: MinDistances, b: MinDistances) {
     return b.minHorizontalDistance - a.minHorizontalDistance;
 }
 // @ts-ignore
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 function compareByMinHorizontalDistanceWithMaxVerticalDistance(maxVerticalDistance: number) {
     return (a: MinDistances, b: MinDistances) => {
         const aIsCloseEnough = a.minVerticalDistance < maxVerticalDistance;
         const bIsCloseEnough = b.minVerticalDistance < maxVerticalDistance;
         if (aIsCloseEnough && bIsCloseEnough) {
             return compareByMinHorizontalDistance(a, b);
-        }
-        else if (aIsCloseEnough) {
+        } else if (aIsCloseEnough) {
             return -1;
-        }
-        else if (bIsCloseEnough) {
+        } else if (bIsCloseEnough) {
             return 1;
-        }
-        else {
+        } else {
             return 0;
         }
-    }
+    };
 }
 function apply(d: MinDistances, loc: SourceLocation): number {
-    if (loc.start === loc.end)
-        return loc.start;
+    if (loc.start === loc.end) return loc.start;
 
     switch (d.horizontalType) {
         case HorizontalClosestDistanceType.LeftIn:
@@ -85,16 +74,20 @@ function apply(d: MinDistances, loc: SourceLocation): number {
 }
 
 export function getDistance(element: HTMLElement, point: Point): ManhattanOffset {
-
-    function distance1D(start: number, end: number, q: number): [number, number] {
-        return [q - start, q - end];
+    function distance1D(start: number, end: number, q: number): { toStart: number; toEnd: number } {
+        return { toStart: q - start, toEnd: q - end };
     }
     const rect = element.getBoundingClientRect();
 
-    const [distanceToLeft, distanceToRight] = distance1D(rect.left, rect.right, point.x);
-    const [distanceToTop, distanceToBottom] = distance1D(rect.top, rect.bottom, point.y);
+    const { toStart: distanceToLeft, toEnd: distanceToRight } = distance1D(rect.left, rect.right, point.x);
+    const { toStart: distanceToTop, toEnd: distanceToBottom } = distance1D(rect.top, rect.bottom, point.y);
 
-    return { offsetFromLeft: distanceToLeft, offsetFromRight: distanceToRight, offsetFromTop: distanceToTop, offsetFromBottom: distanceToBottom };
+    return {
+        offsetFromLeft: distanceToLeft,
+        offsetFromRight: distanceToRight,
+        offsetFromTop: distanceToTop,
+        offsetFromBottom: distanceToBottom,
+    };
 }
 export function getDistance_FOR_TESTING_ONLY(element: HTMLElement, point: Point): ManhattanOffset {
     return getDistance(element, point);
@@ -126,7 +119,6 @@ export function getHtmlElementsWithDataloc(
     isViableToAscend: (element: HTMLElement) => boolean = _ => true,
     isViableToDescend: (element: HTMLElement) => boolean = _ => true
 ): HTMLElement[] {
-
     const result = [];
     for (const element of walkAround(node, isViableToAscend, isViableToDescend)) {
         if (selectLocation(element) !== undefined) {
@@ -134,111 +126,4 @@ export function getHtmlElementsWithDataloc(
         }
     }
     return result;
-}
-
-export type MinDistances = {
-    minDistance: number,
-    minHorizontalDistance: number,
-    minVerticalDistance: number,
-    horizontalType: HorizontalClosestDistanceType,
-    verticalType: VerticalClosestDistanceType,
-};
-export namespace MinDistances {
-    /** Gets whether the point is contained in the rectangle represented by the MinDistances. */
-    export function contains(self: MinDistances) {
-        return VerticalClosestDistanceType.IsIn(self.verticalType) && HorizontalClosestDistanceType.IsIn(self.horizontalType);
-    }
-    /** Gets the manhattan distance from the point to the boundary of the bounding rect represented by the MinDistances. */
-    export function getManhattanDistance(self: MinDistances) {
-
-        if (HorizontalClosestDistanceType.IsIn(self.horizontalType)) {
-            if (VerticalClosestDistanceType.IsIn(self.verticalType)) {
-                return Math.min(self.minHorizontalDistance, self.minVerticalDistance);
-            }
-            return self.minVerticalDistance;
-        }
-        if (VerticalClosestDistanceType.IsIn(self.verticalType)) {
-            return self.minHorizontalDistance;
-        }
-        return self.minHorizontalDistance + self.minVerticalDistance;
-    }
-    export function fromManhattan(q: ManhattanOffset): MinDistances {
-        const minDistance = Math.min(
-            Math.abs(q.offsetFromBottom),
-            Math.abs(q.offsetFromLeft),
-            Math.abs(q.offsetFromRight),
-            Math.abs(q.offsetFromTop)
-        );
-        const minHorizontalDistance = Math.min(Math.abs(q.offsetFromLeft), Math.abs(q.offsetFromRight));
-        const minVerticalDistance = Math.min(Math.abs(q.offsetFromTop), Math.abs(q.offsetFromBottom));
-
-        let horizontal: HorizontalClosestDistanceType;
-        if (Math.abs(q.offsetFromLeft) < Math.abs(q.offsetFromRight)) {
-            // according to the definition of Manhattan."offset", a negative offset means to the left of (both offsetToLeft and offsetToRight), and 0 is in.
-
-            horizontal = q.offsetFromLeft >= 0 ? HorizontalClosestDistanceType.LeftIn : HorizontalClosestDistanceType.LeftOut;
-        }
-        else {
-            horizontal = q.offsetFromRight > 0 ? HorizontalClosestDistanceType.RightOut : HorizontalClosestDistanceType.RightIn;
-        }
-
-        let vertical: VerticalClosestDistanceType;
-        if (Math.abs(q.offsetFromTop) < Math.abs(q.offsetFromBottom)) {
-            // according to the definition of Manhattan."offset", a negative offset means above (both offsetToTop and offsetToBottom), and 0 is in.
-            vertical = q.offsetFromTop >= 0 ? VerticalClosestDistanceType.TopIn : VerticalClosestDistanceType.TopOut;
-        }
-        else {
-            vertical = q.offsetFromBottom > 0 ? VerticalClosestDistanceType.BottomOut : VerticalClosestDistanceType.BottomIn;
-        }
-
-        return { minDistance, horizontalType: horizontal, verticalType: vertical, minHorizontalDistance, minVerticalDistance };
-    }
-
-}
-export enum HorizontalClosestDistanceType {
-    LeftOut,
-    LeftIn,
-    RightIn,
-    RightOut,
-}
-export namespace HorizontalClosestDistanceType {
-    export function IsIn(e: HorizontalClosestDistanceType) {
-        switch (e) {
-            case HorizontalClosestDistanceType.LeftIn:
-            case HorizontalClosestDistanceType.RightIn:
-                return true;
-            case HorizontalClosestDistanceType.LeftOut:
-            case HorizontalClosestDistanceType.RightOut:
-                return false;
-            default:
-                throw new Error("Invalid HorizontalClosestDistanceType");
-        }
-    }
-    export function IsOut(e: HorizontalClosestDistanceType) {
-        return !IsIn(e);
-    }
-
-}
-export enum VerticalClosestDistanceType {
-    TopOut,
-    TopIn,
-    BottomIn,
-    BottomOut,
-}
-export namespace VerticalClosestDistanceType {
-    export function IsIn(e: VerticalClosestDistanceType) {
-        switch (e) {
-            case VerticalClosestDistanceType.TopIn:
-            case VerticalClosestDistanceType.BottomIn:
-                return true;
-            case VerticalClosestDistanceType.TopOut:
-            case VerticalClosestDistanceType.BottomOut:
-                return false;
-            default:
-                throw new Error("Invalid VerticalClosestDistanceType");
-        }
-    }
-    export function IsOut(e: VerticalClosestDistanceType) {
-        return !IsIn(e);
-    }
 }
